HACK: iommu/arm-smmu/qcom: Set SCTLR.HUPCF for everything
Currently this is just set from the special qcom,adreno-smmu iommu.  But
we are not *yet* using that for a5xx.  Meanwhile we still need the HUPCF
bit set so iommu faults in CI do not take out the CP and cause a follow-
on cascade of fail.  On qcom systems, it is safe to set HUPCF everywhere
so do that.

Signed-off-by: Rob Clark <robdclark@chromium.org>