media: Documentation: dt-bindings: ccs: Add module identification
The MIPI CCS specification defines registers for module identification.
However in practice, these registers may be missing in actual
implementations.

Add properties for the module identification instead. Based in this
information the CCS static data is loaded so any other deviances in sensor
read-only registers can be addressed using that instead.

The information may also be used to avoid powering on the sensor during
probe.

Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>