media: ccs: Find sensor CCS static data file based on compatible string
Use the compatible string, as long as it doesn't have "mipi-ccs" or
"nokia,smia" prefix, for loading the CCS static data. This allows using
CCS static data to describe devices that aren't declared as conformant to
CCS but still effectively are such. In this case the vendor may not have
allocated an ID for the sensor so loading the CCS static data file based
on the model (or even vendor) ID isn't an option.

Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>