arm64: dts: qcom: sdm845: Add missing MDSS reset
If the OS does not support recovering the state left by the
bootloader it needs a way to reset display hardware, so that it can
start from a clean state. Add a reference to the relevant reset.

It fixes display init issue appeared recently: without reset
device boots into panel being filled by pixels with different colors
with these messages in kernel log:

[drm:dpu_encoder_frame_done_timeout:2727] [dpu error]enc33 frame done timeout
...
[drm:_dpu_encoder_phys_cmd_handle_ppdone_timeout.isra.0] *ERROR* id:33 pp:0 kickoff timeout 2 cnt 1 koff_cnt 1
[drm:dpu_encoder_phys_cmd_prepare_for_kickoff] *ERROR* failed wait_for_idle: id:33 ret:-110 pp:0
[   90.078161] [drm:dpu_encoder_helper_hw_reset:1763] [dpu error]enc33 ctl 3 reset failure

Fixes: 0e789b491ba0 ("pmdomain: core: Leave powered-on genpds on until sync_state")
Signed-off-by: David Heidelberg <david@ixit.cz>